Fixed db connection error on test environment.
