+++ package.json
@@ -31,7 +31,8 @@
     "supertest": "^3.3.0"
   },
   "jest": {
-    "setupTestFrameworkScriptFile": "./src/config/jest/setup.js",
+    "globalSetup": "./src/config/jest/setup.jest.js",
+    "globalTeardown": "./src/config/jest/teardown.jest.js",
     "verbose": true,
     "automock": false,
     "testEnvironment": "node",
